* possibilité de mettre plusieurs zones dans un seul fichier carte world.xml : zones du monde extérieur caves.xml : cavernes et sous-sols * ajout de ZoneContainer, charge uniquement les zones du ZoneContainer actif * ajout des téléporteurs : possibilité de faire des téléportations intra-zone, inter-zone, et inter-conteneur Tout passage d'un point à un autre est considéré comme un téléporteur (escalier, trou vers un sous-sol, entrée d'un batiment) Quelques démos sur la carte. Les téléporteurs sont sauvegardés en XML dans la zone dans laquelle ils se trouvent. * s'il n'y a pas d'entités ou pas de téléporteurs dans une zone, le tag est optionnel dans le fichier XML (gain de place) * suppresion de tiles redondantes dans le tileset
[éditeur python]
* on peut naviguer dans les zones d'une carte (et les modifier) depuis l'éditeur Python (lecture/écriture)
* support des tags optionnels et des tags des téléporteurs dans le format de carte (mais pas d'interface pour les entités et les téléporteurs pour le moment)

+++ trunk/MediaManager.cpp
@@ -130,9 +130,8 @@ Music* MediaManager::GetMusic(const char* key) const
 }
 
 
-Music* MediaManager::GetMusic(unsigned short id) const
+Music* MediaManager::GetMusic(int id) const
 {
-
 	if(id > musics_.size())
 	{
 		puts("Music: off-bounded index");
